requirements.txt
# Database
supabase==0.7.1
postgrest==0.10.6

# Web Scraping
requests==2.28.2
beautifulsoup4==4.11.2
selenium==4.8.2
webdriver-manager==3.8.5
lxml==4.9.2

# Data Processing
pandas==1.5.3
python-dateutil==2.8.2
uuid==1.30

# Utils
python-dotenv==1.0.0
schedule==1.1.0
pydantic==1.10.7
loguru==0.6.0

# Email Notifications
sendgrid==6.9.7